Initial

+++ 4-2-2021/calculator/calculator.h
@@ -4,6 +4,6 @@
 int addition(int, int);
 int subtraction(int, int);
 int multiplication(int, int);
-double division(double, double);
+int division(int, int);
 
 #endif
